#645 refactored test sources.
