+++ src/index-test.ts
@@ -1,0 +1,6 @@
+import yahooFinance from "./index-common.js";
+import nodeEnvironment from "./env-test.js";
+
+yahooFinance._env = nodeEnvironment;
+
+export default yahooFinance;
