+++ tests-modules/cjs/tests/tests.spec.js
@@ -1,4 +1,4 @@
-const yahooFinance = require("../../../dist/cjs/src/index-node.js").default;
+const yahooFinance = require("../../../dist/cjs/src/index-test.js").default;
 
 describe("esm", () => {
   it("sanity", () => {
