+++ tests/testYf.ts
@@ -1,4 +1,4 @@
-import _env from "../src/env-node";
+import _env from "../src/env-test";
 import _opts from "../src/lib/options";
 import _fetch from "../src/lib/yahooFinanceFetch";
 import _moduleExec from "../src/lib/moduleExec";
